requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
channels>=4.0.0
channels-redis>=4.1.0
requests>=2.31.0
python-decouple>=3.8
pytz>=2023.3
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
django-extensions>=3.2.0
